requirements.txt
fastsam
matplotlib
numpy
Pillow
torch
